+++ Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/EthereumClientIntegrationFixture.cs
@@ -1,13 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
-using System.Reflection;
-using System.Text;
-using Nethereum.XUnitEthereumClients;
+﻿using Nethereum.XUnitEthereumClients;
 using Xunit;
 
-namespace Nethereum.Accounts.IntegrationTests
+namespace Nethereum.Worbooks.Tests
 {
     [CollectionDefinition(EthereumClientIntegrationFixture.ETHEREUM_CLIENT_COLLECTION_DEFAULT)]
     public class EthereumClientFixtureCollection : ICollectionFixture<EthereumClientIntegrationFixture>
