+++ Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/NethereumGettingStartedSmartContractsWorkbookTest.cs
@@ -1,18 +1,19 @@
 ﻿using Microsoft.CodeAnalysis.CSharp.Scripting;
 using Xunit;
 using Nethereum.XUnitEthereumClients;
+using System.Collections.Generic;
 
 namespace Nethereum.Worbooks.Tests
 {
-
     [Collection(EthereumClientIntegrationFixture.ETHEREUM_CLIENT_COLLECTION_DEFAULT)]
-    public class NethereumSendingTransactionsUnitTest : WorbookTest
+    public class NethereumGettingStartedSmartContractsWorkbookTest : WorbookTest
     {
-        public NethereumSendingTransactionsUnitTest() : base(WORKBOOK_PATH)
+        public NethereumGettingStartedSmartContractsWorkbookTest() : base(WORKBOOK_PATH)
         {
         }
 
-        private const string WORKBOOK_PATH = "nethereum-sending-transactions.workbook";
+
+        private const string WORKBOOK_PATH = "nethereum-smartcontrats-gettingstarted.workbook";
 
         [Fact]
         public async void Test()
@@ -20,10 +21,11 @@ namespace Nethereum.Worbooks.Tests
             var code = GetCodeSectionsFromWorkbook();
             //When
             var state = await CSharpScript.RunAsync(code);
-             state = await state.ContinueWithAsync("return (transaction, transactionManagedAccount);");
+             state = await state.ContinueWithAsync("return (transfer.Nonce, balance);");
             var returnValue = (dynamic)state.ReturnValue;
-            Assert.Matches("^0x[0-9a-fA-F]{64}$", returnValue.Item1);
-            Assert.Matches("^0x[0-9a-fA-F]{64}$", returnValue.Item2);
+            //Then
+            Assert.Equal(2, returnValue.Item1);
+            Assert.Equal(100000, returnValue.Item2);
         }
     }
 }
