--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2016-07-28T14-07-00.162110900Z--13f022d72158410433cbd66f5dd8bf6d2d129924
@@ -1,1 +1,0 @@
-{"address":"13f022d72158410433cbd66f5dd8bf6d2d129924","crypto":{"cipher":"aes-128-ctr","ciphertext":"3d4bf46f682b50209353dbf915d54b3f9ec5a83dd5b861ccb540820ba3f092c6","cipherparams":{"iv":"777397eec0065be6879e2591dc8ad6d6"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"2daaf9451ad16e6bfb45cafcfd8626dc6663a54e638147525bd537c6d1e08a92"},"mac":"02aad6b00c98d1223b284934065a8ec90d7e66796dbd9fafc50869fd018038d1"},"id":"13beb70e-1c95-4bb1-b469-974717b0036d","version":3}
