--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-06T17-00-58.642215000Z--b972ce2413ac102f650d3cea5201739e5f6c9022
@@ -1,1 +1,0 @@
-{"address":"b972ce2413ac102f650d3cea5201739e5f6c9022","crypto":{"cipher":"aes-128-ctr","ciphertext":"41f2d3ba61ef65a78a4b1e1e2044c8d6bd6559af915a6dfbae559d53132f16e4","cipherparams":{"iv":"7ec532a570729fe909f8960cbf276f0d"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"367dde68d4a8128c1aaecd14783602b38691063e75f5d9808e50ca370c95facc"},"mac":"c5a2dc969c3e612512f2a73e0f343d644934cbbb2355286eedec446d65925b7e"},"id":"d9945366-ab9b-4432-873d-ca881fef69b7","version":3}
