--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-06T17-01-21.273130500Z--e229dd32d1720aa1b3cbf6ce8a4409c1bb9e8278
@@ -1,1 +1,0 @@
-{"address":"e229dd32d1720aa1b3cbf6ce8a4409c1bb9e8278","crypto":{"cipher":"aes-128-ctr","ciphertext":"81318fbd26e893765cac3bc41751602f29795939cb4f4bc6604167222d077b8b","cipherparams":{"iv":"68c911f143b5f61e4ebc60abed73882b"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"60223ee5a6f6ac2c2f21dbcf43c1b7b62e084912d067adb4707bdbca08200d65"},"mac":"b36fe5c4a7731bf704d32332785a052b73d8fb47905b00bd1540bb923ed02811"},"id":"a6544ce2-b097-4c19-9633-6ec40faa8e0e","version":3}
