--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-06T17-01-23.115868000Z--beb827aa4ec1ccf9c0f617a3be1bb66d2203e7c9
@@ -1,1 +1,0 @@
-{"address":"beb827aa4ec1ccf9c0f617a3be1bb66d2203e7c9","crypto":{"cipher":"aes-128-ctr","ciphertext":"3848d5dd32b85bbcb2c5ffc5d9034c4f680be0d726c337351bbe4aa593aaaa40","cipherparams":{"iv":"97e9091d67ffe6646c7a0be4416927da"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"4b9844db90ecb69bcf8f289e5b8eb18a2dbe669f6f47a6a2d462e6b79d7eabd3"},"mac":"a2a826714f85471725129934a1f6e9ae14ddd6606ece98fc17bb2c7f3455a606"},"id":"a87f6bb5-5064-42e9-aef2-942cf46eec12","version":3}
