--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-06T17-01-26.195938600Z--76846c1d0cd69345f547d36867a2b1c33a26de41
@@ -1,1 +1,0 @@
-{"address":"76846c1d0cd69345f547d36867a2b1c33a26de41","crypto":{"cipher":"aes-128-ctr","ciphertext":"10834c5b58d575273ee26d2a02ba54ebe0f44a05fbc6c08759d356db8eb1d267","cipherparams":{"iv":"18cd4be220831054f1208b6038c213ea"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"b6edf705762123d1dd19b632374b6ffdf7a470d2f23d62c71663c5222ca5c17e"},"mac":"0e572bc79ddc46a4bd6cca00e22d220fb657427d80c43182abbafed14a02b86e"},"id":"404e5c53-027b-4c43-aac5-29eb6fe555f7","version":3}
