--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-06T18-11-07.633814100Z--64f8ed538c8b37066593f26ac3f88bb7de7f9346
@@ -1,1 +1,0 @@
-{"address":"64f8ed538c8b37066593f26ac3f88bb7de7f9346","crypto":{"cipher":"aes-128-ctr","ciphertext":"22a81f9175f312691c1951d265b80004f18cf46adfd212beb27f433d15fbb83d","cipherparams":{"iv":"493fae8288b4c366614aeb9e5d2751fc"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"3f7814c556360a85014f7525db01c9bab0f5bac8159b486726d967e9b2a18fc4"},"mac":"2d559675497c4148e7fe8d30106c1487280abfe61f814334e7ba071606a7aac1"},"id":"af3eb035-de51-4c9d-a475-206ccb0ae49d","version":3}
