--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-13T23-26-29.935709800Z--0cbf847abc15d85662e7717299bd53b7675fc14b
@@ -1,1 +1,0 @@
-{"address":"0cbf847abc15d85662e7717299bd53b7675fc14b","crypto":{"cipher":"aes-128-ctr","ciphertext":"212bbb0c665119dc8fd2890a4d70e3fee7724ac18d052e51c108fe3641a5f37f","cipherparams":{"iv":"2059ea3ee8303f68652e99c168c573be"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"e343535e7fb2d22758ec051986402944973376d49e5283f509e085697caeef38"},"mac":"34252aff5ca8d5d8ec9d005248cbafdda7e73d9e93fffd3160e7ac95b0341d52"},"id":"60cf75b0-6d00-448a-92be-210a882d9a0e","version":3}
