--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-22T21-31-48.813069500Z--2411837ed6c9084c1a84d4f3e43306af20af9745
@@ -1,1 +1,0 @@
-{"address":"2411837ed6c9084c1a84d4f3e43306af20af9745","crypto":{"cipher":"aes-128-ctr","ciphertext":"b0a3da9dc23b9f1f7681ec2903d234c690acb37ec89d9d81ab1f8791bd96dd36","cipherparams":{"iv":"a1583623044b12e290e9f2e2ed38059f"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"25319c6027a8a7447ea4f815e3de44c1272dedf892006cba1da67e35c499381c"},"mac":"72279fca1ed8f81058e1cfe98c59c3446e7ed9530b3fa755fb00d1b769330877"},"id":"982739a3-9407-4ebc-ad40-78e002c81af3","version":3}
