--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-22T21-32-10.250654100Z--433daeec3623c4f571bc4056920d6f8108fa9b3e
@@ -1,1 +1,0 @@
-{"address":"433daeec3623c4f571bc4056920d6f8108fa9b3e","crypto":{"cipher":"aes-128-ctr","ciphertext":"c0d33002a8b01a7e0ce9f542a90a239421e0e5cf1826708a64364daa0c2b9913","cipherparams":{"iv":"6f620d1bc38bd1efb225d270a015cf2c"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"c63620616b6d6107678accc0a73b76ad9d78b6a952db110e73fc1e6bc14397e5"},"mac":"ce1f18f6094a366b0b5522af0369e1d8161e8248571a116186f13a1a9577428b"},"id":"b584479a-ad3f-4baf-985f-6f9759bd7439","version":3}
