--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-22T22-08-54.257848500Z--18ea15c599091c41d73fa3f1ebbad45babd0f7dd
@@ -1,1 +1,0 @@
-{"address":"18ea15c599091c41d73fa3f1ebbad45babd0f7dd","crypto":{"cipher":"aes-128-ctr","ciphertext":"fc7b57fb308134bf7e5aee49a59b3ec04659e6e2497f21cbd1b1d517d40026f6","cipherparams":{"iv":"660ea4eb6e94d9710c19d00c8b66e7ae"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"400e807e87528768c692fdd080be4eb85b18aeeba1a9175f6d48c5bb55209d80"},"mac":"9e1a979637c31359240717d8a01d8be61657959ef49d5f8fd71a2790d83ec816"},"id":"de3335b9-2f45-46ff-a348-76a9d095866f","version":3}
