--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-23T14-59-06.497222400Z--b76a9b2d940b2018b1a559d5fc6d0c66bfb07fb3
@@ -1,1 +1,0 @@
-{"address":"b76a9b2d940b2018b1a559d5fc6d0c66bfb07fb3","crypto":{"cipher":"aes-128-ctr","ciphertext":"52df786668cccd4756b48b74be3323b30863c07def9c4fb2aa5a038ee15a2c1b","cipherparams":{"iv":"8a147e53e082b656f981d34d16de64ee"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"499b10db0238eb144877e277f8f631937c8a5e1470ea3f5882ff03d784ec6826"},"mac":"e9ab37b84ede0eeec32e3c2d8169df3e7606b8487bbfb4633bdf92d1261c54c0"},"id":"c455a05f-aa8c-448f-8987-e2d598b8665b","version":3}
