--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-23T15-04-59.153492900Z--be4c14a4a3300b6518a4a9ba72839df415f2d5fc
@@ -1,1 +1,0 @@
-{"address":"be4c14a4a3300b6518a4a9ba72839df415f2d5fc","crypto":{"cipher":"aes-128-ctr","ciphertext":"3cf9c2d0cbd68c22b7ddb827b383978884a67f6e45f243f81f3d9221130ca026","cipherparams":{"iv":"af4f937f22e1c85274e473a2ef460dc4"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"308c1c7592b5673d3783fe047139d15a5af8653af8451b7eb950b952d9afd2f4"},"mac":"874f9d9910221881a311fcd2357e6c8f5942b8db3dc6aa32db226f88444fb7af"},"id":"e742a9be-bcd4-4d1f-8acc-cf9e1bfb8cee","version":3}
