--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-23T15-26-31.106706100Z--f37731d6da41c87fa2b90508daeb2d525d1ca76d
@@ -1,1 +1,0 @@
-{"address":"f37731d6da41c87fa2b90508daeb2d525d1ca76d","crypto":{"cipher":"aes-128-ctr","ciphertext":"410c859bf4d0019326da5126c83144ddcbd15ea596ffcc6fb1e0b143b698b509","cipherparams":{"iv":"3e4331758e1208943ab16a079bc32164"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"62213f683acb3777ca6179cd2798319dd57c7b50064b8b8bb7fdf0dfdd18f0cf"},"mac":"ce1ae58d6c9a4c1e01508371d9358f84839da29f58130199b098b4937ff8bbc1"},"id":"6a45490f-1646-4461-8cda-e2e810c6419f","version":3}
