--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-23T15-36-29.452085200Z--fb1bac8d62cc40b04d203aff986770adcff61c13
@@ -1,1 +1,0 @@
-{"address":"fb1bac8d62cc40b04d203aff986770adcff61c13","crypto":{"cipher":"aes-128-ctr","ciphertext":"7bb2974ef9d8862b4a2a1305159703028fca29be9fd4ba927144c59351dc406b","cipherparams":{"iv":"fd1f6d5c5c5949387e789b8f132a7d9c"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"f515cc4bba5a9bb35aaaf3a086018b2a23ffa00172aaa9e254ec9a1aaded1a43"},"mac":"fab59a280e8fae6a4de25036e90f68d200babfbf7ce600a037ff044c5456373f"},"id":"a87786cd-c69d-4727-8f6f-708889e377e8","version":3}
