--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-23T15-52-46.251991800Z--d0da5d767581d0e4f2542879f1ef3737ef2f9e82
@@ -1,1 +1,0 @@
-{"address":"d0da5d767581d0e4f2542879f1ef3737ef2f9e82","crypto":{"cipher":"aes-128-ctr","ciphertext":"0d478627683559ca96075bbc0e2c5a25e8103f0a98ee03b249d0e4e5f5186885","cipherparams":{"iv":"0deb2354b5ceed786683df9368e80f9f"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"11ed7204e80518bf21b9d935b14ad3004a21a7c2eb54285626044a9a9ddc7e52"},"mac":"351efc36f71bb8d77b3cb65fc4332743f2df181b19360a1417db40a704f10c62"},"id":"65d2ad19-c1c5-4b78-9f7a-3f398deebab3","version":3}
