--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-23T15-58-42.705149700Z--5a5795a721bd96968e895502d48a4a203c000442
@@ -1,1 +1,0 @@
-{"address":"5a5795a721bd96968e895502d48a4a203c000442","crypto":{"cipher":"aes-128-ctr","ciphertext":"6fd7c5214a889ba72edc2ede0408feaa7e8032afb3230458a71ce7a45b8e5745","cipherparams":{"iv":"818f3be6f4cf30272e63e451fbdbc007"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"21972aef57ac3305a036260ef7ec50e2a7df78e33f3ec4c95915e0b30a6768c0"},"mac":"0f320d6f60db5a4e174d7f14e4d878940302f8fe925413fb3118ee557f25006e"},"id":"3b34f277-5e28-425c-ae8b-3b15a85e1a7d","version":3}
