--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-23T16-29-00.082144200Z--628671cd929551ae8a4d95f458998d1f805586b4
@@ -1,1 +1,0 @@
-{"address":"628671cd929551ae8a4d95f458998d1f805586b4","crypto":{"cipher":"aes-128-ctr","ciphertext":"161affa348e0cc2dea257932085bd8daaf2098d669dd8426915f46380ed74a12","cipherparams":{"iv":"a2df7a1350c98ef9d25a60380d868643"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"0cce4ef0229e4e73b6fcfd5bf18ce1784ab08a276fe28f1374dc622ac57a2fe9"},"mac":"8202861145b097c22877a64452db96a59486086310278e2d14c80b39c8169c49"},"id":"72fd6b47-d517-4e13-8b6f-96b199f76650","version":3}
