--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-23T16-33-46.672676000Z--aa1f999becbde2b344b5104cb2caf99a778b1d16
@@ -1,1 +1,0 @@
-{"address":"aa1f999becbde2b344b5104cb2caf99a778b1d16","crypto":{"cipher":"aes-128-ctr","ciphertext":"f3b24fbd8872d0fe930598d90b5c4cfeda7bf2987eac145e1443eb5f02e88001","cipherparams":{"iv":"4ecc01381d551ee674f460699be1f433"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"a99b78777f737fb3ead4a3fbc00b0c29289e4ef2db982be8027252c34643a59c"},"mac":"8fe724b2a1176f4ff036a4273715712e105fca0c7f67293157b2cae0f30432bf"},"id":"2438577b-52c3-47b0-a27b-a706284bc826","version":3}
