--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-23T16-42-16.290014200Z--2320d63a85d51ed77a8bc8d121987220ef5c57f0
@@ -1,1 +1,0 @@
-{"address":"2320d63a85d51ed77a8bc8d121987220ef5c57f0","crypto":{"cipher":"aes-128-ctr","ciphertext":"dd0260b7e26e9877a005cf1ddf146d0144824266a15818ebf2770bce21ebdbb3","cipherparams":{"iv":"70f4f062786a97f7ba3dc28130729c60"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"f24a3d7ebf5c94db313d67490f41f2582fbf6a60053f30a8f576b2c1e9e57fc8"},"mac":"26a9c1d927b553559cadd8c751346b33a5ef400cd8ad03592c80b1e6de0ead9a"},"id":"cd42eb18-53cf-48fe-a458-152b3accafec","version":3}
