--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-23T17-14-15.474573700Z--75b5b6ce013707c5e097b5507d59c6ba039e1a33
@@ -1,1 +1,0 @@
-{"address":"75b5b6ce013707c5e097b5507d59c6ba039e1a33","crypto":{"cipher":"aes-128-ctr","ciphertext":"ae8f8d7a08b077e885b111e5cee5627e68e0c3c52258f39546df6e9d6a3c54c8","cipherparams":{"iv":"9c37686c1feb222b587def0a03541d47"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"ce9effe2d6954e74f71f005a3a2d4b7c04af3e86fbec3082b7908ea4ea351938"},"mac":"d92233003ec301db7a1cfce099690a59ad7337b08c634e3793776dc6380ae636"},"id":"37a33321-6340-4c6c-806b-48f888234cca","version":3}
