--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-08-23T17-55-44.278359800Z--659a07dfa80e7570ba9e2746606c9fcc346e15c8
@@ -1,1 +1,0 @@
-{"address":"659a07dfa80e7570ba9e2746606c9fcc346e15c8","crypto":{"cipher":"aes-128-ctr","ciphertext":"b90b26ecaf963fd2ffe516591b7cd453a7bcb0f63a75c43871d15dcb844dabf4","cipherparams":{"iv":"70a9de5a2bdb8e84e81be239e470aa9c"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"7400b24e56af23d7652f262528045a55ce5db09724be74cd5d14c84e78ab1a9f"},"mac":"3e0cd60010bae3d38d9ef3ea048bdb3baf96520e6992e4c8ecf82b13ca78430e"},"id":"edfe185f-e90e-4be0-b30d-97a00cf311c1","version":3}
