--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-12-09T18-00-11.185428100Z--9365b990202e0f2ea77084d0b8f7e32db229a564
@@ -1,1 +1,0 @@
-{"address":"9365b990202e0f2ea77084d0b8f7e32db229a564","crypto":{"cipher":"aes-128-ctr","ciphertext":"20228e374d057de66980c7df4f4a2417172988a0458023dc14cb6b766251d835","cipherparams":{"iv":"7ecbc8c2a1eaf1e5cab36e28e4f3ed53"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"64a1319a441ceef59dc824917e6688e6fd43eab09f7bfa47365e7dbb37220680"},"mac":"6b87375d2b9b087fec36aa63100e929cbd58662b4b1c7a405bcfdaea3312cf2c"},"id":"d78e477c-2d78-4885-a251-618b1927f093","version":3}
