--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-12-09T18-29-57.006840800Z--d38b915987b049fdf100889297d876526ad93bec
@@ -1,1 +1,0 @@
-{"address":"d38b915987b049fdf100889297d876526ad93bec","crypto":{"cipher":"aes-128-ctr","ciphertext":"72f8c6674c452340faf10bb9948ededfbf575b30542c5e4c407da1229773d4f5","cipherparams":{"iv":"8e809b0151f40311bfef016db89363f9"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"bc1f00c79061a3e56ca04359136653bbc4b6b51d13812dd8866a8e76163e0825"},"mac":"6dbb0fe7067218ccd83b035730c074a8b34a084b0439095583242ce7e73a2bb8"},"id":"ada4a57c-4888-4b3b-bd29-9eab4adc99fd","version":3}
