--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-12-09T18-31-01.348591100Z--904af442cda9c46a0a21805a7f13b34f4dafa015
@@ -1,1 +1,0 @@
-{"address":"904af442cda9c46a0a21805a7f13b34f4dafa015","crypto":{"cipher":"aes-128-ctr","ciphertext":"9f087c6a42a36f1e29a69ae7bb11923d4f5c74d4c6cdb6f62b7634cfb19e2124","cipherparams":{"iv":"dc98d60a3e174495bbc4d6517bbbf66c"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"4c68e2de3f8b3d21ff19d2587f8db3fbd3d2d729d4cada23474b1bb55c046f50"},"mac":"2a7444cd3dc5019109b6f405c143026aa42be9b32fbea0088599d4c5bbf7289b"},"id":"981e0a07-bcd9-484b-9194-8e3cb6b54c4e","version":3}
