--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-12-12T17-24-44.133367100Z--4f749e03bb5065a417be5e179abf8fbdcc4481cc
@@ -1,1 +1,0 @@
-{"address":"4f749e03bb5065a417be5e179abf8fbdcc4481cc","crypto":{"cipher":"aes-128-ctr","ciphertext":"afe3994248d3960056d69f85e46306e4120364ceb010d4cc67ce79344b5efcd1","cipherparams":{"iv":"8f8ad277fe36fdee1646f708464e8518"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"09c62d9f91ee7645b9012eb8888178040e34f86a3a8aceab08ab46fd85bd0409"},"mac":"749de4b565edb1a63b37e7f5665ba9057dc005176a3aa7ee37c4a6a83ae70e08"},"id":"bbc2008e-b92e-49c9-b000-9e237570169b","version":3}
