--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-12-12T17-28-38.631211100Z--3a75d31db0e6919e0d292e22153a6613f98824e3
@@ -1,1 +1,0 @@
-{"address":"3a75d31db0e6919e0d292e22153a6613f98824e3","crypto":{"cipher":"aes-128-ctr","ciphertext":"e03ae5e23e7faed59f2178b38c03a6509fa46176950230edd38ec2d28d56aaf5","cipherparams":{"iv":"25be6c986667842d875c64037519b0ee"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"d3d43760ea2f9d341e00596ce45d6f040203ed9f898bd7c517975b02e585edae"},"mac":"de5341ae97dcc76ab2b458d26ad91f855d92ae51340a58d819b4ae8a040a1011"},"id":"7780b4b2-b284-4efb-9752-c38e66091e89","version":3}
