--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-12-12T17-30-05.574513500Z--f3ac6ce755052086757aa1cb80b10e6a8217da58
@@ -1,1 +1,0 @@
-{"address":"f3ac6ce755052086757aa1cb80b10e6a8217da58","crypto":{"cipher":"aes-128-ctr","ciphertext":"9afb941ad848a3b876acdb29eec00613560120cd88c78da7579d37c803ec49fd","cipherparams":{"iv":"a73e82709e9f403b3fa35ead4555fb34"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"6e515ffa3a5c0e7a4892d1d8dee0a3e23eb7970abba86d7e22375b87e7d2514f"},"mac":"7eccd6c4ec64db92a98fc8402e1d074d3feb4c4486753dfa5e3ac08ac6356a58"},"id":"3c76843f-c773-4fb0-a69d-2c5159b2a232","version":3}
