--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-12-12T17-34-55.048776100Z--b023727b583c772fb812ad9c2842d5c79a2ecffe
@@ -1,1 +1,0 @@
-{"address":"b023727b583c772fb812ad9c2842d5c79a2ecffe","crypto":{"cipher":"aes-128-ctr","ciphertext":"fe486536f043de14bf8168cbf50ee068add10ce08d74a8b6678a4566acadbe82","cipherparams":{"iv":"7acbc3e81fa954a65f4ff5c1d3010ce6"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"a75d3b6e66a9cc1efb6bb56fd06d8979b31836fcf8f9ce34a53fd7dc66869c72"},"mac":"0ae3a204b59486d3ffd42139a27c39ea41f6cfc45545125407acc99830f5224d"},"id":"818dc6ae-7157-40b3-964e-c143753f41d7","version":3}
