--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-12-12T18-06-14.357631800Z--7310aa669b1a607045b36650a812a6f61abe5e21
@@ -1,1 +1,0 @@
-{"address":"7310aa669b1a607045b36650a812a6f61abe5e21","crypto":{"cipher":"aes-128-ctr","ciphertext":"40f11e9af477c1f479dcd28adc730d9daf375a61836406b341faae95e45dd4b3","cipherparams":{"iv":"3a526b040d8dba54fb16833ccb6b0ec0"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"f05ba5534585826b1a68e827f2063453e4ff1deca7fc29c99df80594ebfcdaa2"},"mac":"3ea93ae6361077be14d41894cc9f7e752cadb6152591589e0148721a2f92d28a"},"id":"c9cf716d-490d-48b2-bc87-8bf5940f2d12","version":3}
