--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-12-12T18-11-31.238847800Z--ec849a65639a9240068f1f663e36608e8bf87538
@@ -1,1 +1,0 @@
-{"address":"ec849a65639a9240068f1f663e36608e8bf87538","crypto":{"cipher":"aes-128-ctr","ciphertext":"aef895ab4fdb3c3e11ba0ce5bcbd113e49d066dc4934afd1c00987a878c59c55","cipherparams":{"iv":"f293ed104e53bd6ffa064d36633d116a"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"1cc437586eab98324df2771388ae15b969eb1f629fae7d7f300ee99780494b28"},"mac":"00753ec25f2bbfa52ed05c7c05709d697f34047debdc536857307dd87d34423d"},"id":"1e36135e-25c8-42e8-9f94-e4fb2c05a758","version":3}
