--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-12-12T18-20-18.569989300Z--4466083588c4e5e5bdb6cfb53c0c8818448189ed
@@ -1,1 +1,0 @@
-{"address":"4466083588c4e5e5bdb6cfb53c0c8818448189ed","crypto":{"cipher":"aes-128-ctr","ciphertext":"1dd19b05e05bcd563d3b9dd6ebc555fd536c3a210c7c01fe5d1e7dc6c43f30b0","cipherparams":{"iv":"32b3e72587fb6df48857c9afa75fe35d"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"4a6a3fe3a8e4bd7f2a5a6541127741b70eede70c1d104372d2ad46f2f0710fdb"},"mac":"2d5e3a65a2b8b8d72d5019b405ee48ed763ea4c407465b530ba4e08f55e98f36"},"id":"d3805fd6-11b7-4f1e-9204-c2ce168b8f4a","version":3}
