--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/keystore/UTC--2018-12-12T18-26-29.744676800Z--2715ed441b89fd3b02302c04e1e0e083f9f659a8
@@ -1,1 +1,0 @@
-{"address":"2715ed441b89fd3b02302c04e1e0e083f9f659a8","crypto":{"cipher":"aes-128-ctr","ciphertext":"a330559fd8bb234a5a399d2a12af27f2a9a595d7c41a0816ffb55ebf463dac84","cipherparams":{"iv":"6211311d3c030a13056a5f36d6a9a8a1"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"b3c8058b07f788feea9bc34dd524b166ad277a3a6e9af2d3c89b752d9bc85521"},"mac":"04a8b9d7297b5c1f295a99625b23699a1ba970e4796ad0909251b99784c7ac16"},"id":"44253380-9e65-44ac-959a-1e26479332f8","version":3}
