--- Nethereum.Worbooks.Tests/Nethereum.Worbooks.Tests/nethereum-creating-a-new-account-using-geth.workbook
@@ -1,46 +1,0 @@
----
-uti: com.xamarin.workbook
-id: 5ae81721-2480-4fb0-baa6-62e1f79aa238
-title: nethereum-creating-a-new-account-using-geth
-platforms:
-- Console
-packages:
-- id: Nethereum.Web3
-  version: 3.0.0
----
-
-## Creating a new Account using Geth Personal Api
-
-First we need to add a package reference to Nethereum.Web3. This package includes the base Personal management APIs as it is standard across clients.
-
-```csharp
-#r "Nethereum.Web3"
-```
-
-Documentation about Nethereum can be found at: <https://nethereum.readthedocs.io>
-
-## Prerequisites:
-
-First, let's download the test chain matching your environment from <https://github.com/Nethereum/Testchains>
-
-Start a Geth chain (geth-clique-linux\\, geth-clique-windows\\ or geth-clique-mac\\) using **startgeth.bat** (Windows) or **startgeth.sh** (Mac/Linux). The chain is setup with the Proof of Authority consensus and will start the mining process immediately.
-
-We can now create a new instance of Web3:
-
-```csharp
-#r "Nethereum.Web3"
-```
-
-```csharp
-using Nethereum.Web3;
-```
-
-```csharp
-var web3 = new Web3();
-```
-
-Using the Personal API we can now execute the request "NewAccount" with a password to encrypt the account storage file:
-
-```csharp
-var account = await web3.Personal.NewAccount.SendRequestAsync("password");
-```
