+++ testchain/clique/devChain/keystore/UTC--2018-12-20T07-04-14.348958700Z--e47af028d5ac4d3396e9ca29933b612396d61879
@@ -1,0 +1,1 @@
+{"address":"e47af028d5ac4d3396e9ca29933b612396d61879","crypto":{"cipher":"aes-128-ctr","ciphertext":"360db685569179d13f17029f90c8eb86dfa41c6f999b84784486e4edb278eb54","cipherparams":{"iv":"452e68ad0be5258f4f0bf94521f625c6"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"d171e97545f0a540a26610cac2d39398bc5c85eddab47867e9b5bd5bf08276e9"},"mac":"a7831ec0665412c6a39cf0b7f0763ee96fb0f70a66ac888c69866297531a1b61"},"id":"568a9213-a65a-4625-b1c2-290694e1bf66","version":3}
