+++ testchain/clique/devChain/keystore/UTC--2018-12-20T07-06-33.857486100Z--e82a3eb6d0ac76783314158f19da278c2d1f43e0
@@ -1,0 +1,1 @@
+{"address":"e82a3eb6d0ac76783314158f19da278c2d1f43e0","crypto":{"cipher":"aes-128-ctr","ciphertext":"bb20316eb098ab9d0fc807fc651f9a1780df87f1c1411767b8af4584141e2fd9","cipherparams":{"iv":"a1cc8163d8ffb0bbd9e33ce6a87f9190"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"847560a8db96a3378a012d976f2c502666ad67acb0be23c6ee938437cb32a0ac"},"mac":"9b3797c5b8eefe15255f573df388548af844fe62b9ab5a61e555e57ca9736ee9"},"id":"cc6c1b43-23f8-4df8-9af2-306036485800","version":3}
