+++ testchain/clique/devChain/keystore/UTC--2018-12-20T07-16-35.436674200Z--77696a988cc643581061831377e68f64031b7df1
@@ -1,0 +1,1 @@
+{"address":"77696a988cc643581061831377e68f64031b7df1","crypto":{"cipher":"aes-128-ctr","ciphertext":"95de251554f81e3435f00af68f3937e4b5edf4df2ddab5c66e8bf3df4bad1597","cipherparams":{"iv":"b818f157b256235f46ba1a8af5edfa90"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"84a9005812f67db02dfd1388eef3e0b5ee98601657188a535fc9ea7f36c7660d"},"mac":"3ab05c18891da07d1758f8364ce52837b6ec766716e0f79bc3c8bcd449cca735"},"id":"e79fc3db-7a9f-4359-9a1d-4d9b1499091e","version":3}
