+++ testchain/clique/devChain/keystore/UTC--2018-12-20T07-22-39.109125200Z--a047e71c45c4c67d51c5b8e1eea2e7b905e22fd3
@@ -1,0 +1,1 @@
+{"address":"a047e71c45c4c67d51c5b8e1eea2e7b905e22fd3","crypto":{"cipher":"aes-128-ctr","ciphertext":"f127e09f84a9722beda7e8dcbce96994c5e866599b571109d280db93b3fd8c62","cipherparams":{"iv":"c61c9b5e3429a6d4e594fad2f298fa94"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"92b7cda3f0ed365d97e685f000a2c55d231c9727bc6f14053190c8d9aee4c90a"},"mac":"97f46aa8bcee4dc29117b86c166e5ff5df70159bdf9e45faf81c3c41033cca89"},"id":"ce1af4b6-8d71-42b2-afb3-6533fd2c620f","version":3}
