+++ testchain/clique/devChain/keystore/UTC--2018-12-20T07-45-08.028509200Z--a78339e57e8cba2e5b90213d97b6afeb4085b8c0
@@ -1,0 +1,1 @@
+{"address":"a78339e57e8cba2e5b90213d97b6afeb4085b8c0","crypto":{"cipher":"aes-128-ctr","ciphertext":"bf035d8998e432c68ef41fedabebc439e73a417bab85584f79f5bc6341f2578c","cipherparams":{"iv":"b65e05d4b700b65a99d923134ab654a3"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"d2b710972b854e77fc2733eb63fa7dc5fcd4911901cfa2125438388f8de58d98"},"mac":"34c922afd29a0bdfb35d2111c1b4d9c10db9352c96c32c37ad5ffd62c8c7f7d1"},"id":"f61fbc5f-da78-4d81-831d-eef16bd7f2a1","version":3}
